requirements.txt
fastapi==0.129.0
motor==3.7.1
pymongo[srv]==4.16.0
python-dotenv==1.2.1
uvicorn==0.41.0
certifi==2026.1.4